Move functions to file if not used in multiple places
